requirements.txt
# Confluence integration
atlassian-python-api==3.41.0
beautifulsoup4==4.12.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0
pgvector==0.2.4

# AI/ML - Updated versions
anthropic>=0.25.0  # Latest version
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
sentence-transformers==3.0.1
transformers>=4.44.0
huggingface-hub>=0.24.0
torch>=2.0.0
safetensors>=0.4.0
openai>=1.10.0

# Utils
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
pandas==2.1.4
numpy<2.0.0
httpx  # Required by newer anthropic versions

# Caching
redis==5.0.1